+++ tests/ops/api/v1/endpoints/test_privacy_notice_endpoints.py
@@ -43,6 +43,7 @@ class TestValidateDataUses:
             consent_mechanism=ConsentMechanism.opt_in,
             data_uses=["placeholder"],
             enforcement_level=EnforcementLevel.system_wide,
+            displayed_in_overlay=True,
         )
 
     @pytest.fixture(scope="function")
@@ -494,6 +495,7 @@ class TestGetPrivacyNotices:
                         "consent_mechanism": ConsentMechanism.opt_in,
                         "data_uses": ["advertising"],
                         "enforcement_level": EnforcementLevel.system_wide,
+                        "displayed_in_overlay": True,
                     },
                 )
             )
@@ -539,6 +541,7 @@ class TestGetPrivacyNotices:
                         "consent_mechanism": ConsentMechanism.opt_in,
                         "data_uses": ["advertising"],
                         "enforcement_level": EnforcementLevel.system_wide,
+                        "displayed_in_overlay": True,
                     },
                 )
             )
@@ -690,6 +693,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     )
                 ],
                 {
@@ -708,6 +714,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         )
                     ],
                 },
@@ -728,6 +737,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                     PrivacyNotice(
                         id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -741,6 +753,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                 ],
                 {
@@ -759,6 +774,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         ),
                         PrivacyNoticeResponse(
                             id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -773,6 +791,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         ),
                     ],
                 },
@@ -793,6 +814,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                     PrivacyNotice(
                         id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -806,6 +830,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                 ],
                 {
@@ -824,6 +851,7 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
                         ),
                         PrivacyNoticeResponse(
                             id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -838,6 +866,7 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
                         ),
                     ],
                     "third_party_sharing": [
@@ -855,6 +884,7 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
                         ),
                     ],
                 },
@@ -876,6 +906,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                     PrivacyNotice(
                         id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -889,6 +922,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                 ],
                 {
@@ -906,6 +942,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         ),
                     ],
                     "third_party_sharing": [],
@@ -926,6 +965,9 @@ class TestGetPrivacyNoticesByDataUse:
                             "third_party_sharing"
                         ],  # data use is not applicable to any system
                         enforcement_level=EnforcementLevel.system_wide,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                 ],
                 {"advertising": []},
@@ -947,6 +989,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                     PrivacyNotice(
                         id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -960,6 +1005,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                     PrivacyNotice(
                         id=f"{PRIVACY_NOTICE_NAME}-3",
@@ -973,6 +1021,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                     PrivacyNotice(
                         id=f"{PRIVACY_NOTICE_NAME}-4",
@@ -986,6 +1037,9 @@ class TestGetPrivacyNoticesByDataUse:
                         created_at=NOW,
                         updated_at=NOW,
                         version=1.0,
+                        displayed_in_overlay=True,
+                        displayed_in_privacy_center=False,
+                        displayed_in_api=False,
                     ),
                 ],
                 {
@@ -1004,6 +1058,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         )
                     ],
                     "provide.service.operations.support.optimization": [
@@ -1022,6 +1079,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         ),
                         PrivacyNoticeResponse(
                             id=f"{PRIVACY_NOTICE_NAME}-3",
@@ -1036,6 +1096,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         ),
                         PrivacyNoticeResponse(
                             id=f"{PRIVACY_NOTICE_NAME}-2",
@@ -1050,6 +1113,9 @@ class TestGetPrivacyNoticesByDataUse:
                             updated_at=NOW,
                             version=1.0,
                             privacy_notice_history_id="placeholder_id",
+                            displayed_in_overlay=True,
+                            displayed_in_privacy_center=False,
+                            displayed_in_api=False,
                         ),
                     ],
                 },
@@ -1121,6 +1187,7 @@ class TestPostPrivacyNotices:
             "consent_mechanism": ConsentMechanism.opt_in.value,
             "data_uses": ["advertising"],
             "enforcement_level": EnforcementLevel.system_wide.value,
+            "displayed_in_overlay": True,
         }
 
     @pytest.fixture(scope="function")
@@ -1132,6 +1199,7 @@ class TestPostPrivacyNotices:
             "consent_mechanism": ConsentMechanism.opt_in.value,
             "data_uses": ["personalize"],
             "enforcement_level": EnforcementLevel.system_wide.value,
+            "displayed_in_overlay": True,
         }
 
     def test_post_privacy_notice_unauthenticated(self, url, api_client):
@@ -1365,6 +1433,74 @@ class TestPostPrivacyNotices:
         )
         assert resp.status_code == 200
 
+    def test_post_privacy_notice_opt_in_must_be_displayed_in_overlay(
+        self,
+        api_client: TestClient,
+        generate_auth_header,
+        notice_request: dict[str, Any],
+        url,
+    ):
+        """
+        Test opt in consent mechanisms must be displayed in overlay at minimum
+        """
+        assert notice_request["consent_mechanism"] == ConsentMechanism.opt_in.value
+        notice_request["displayed_in_overlay"] = False
+
+        auth_header = generate_auth_header(scopes=[scopes.PRIVACY_NOTICE_CREATE])
+
+        resp = api_client.post(url, headers=auth_header, json=[notice_request])
+        assert resp.status_code == 422
+        assert (
+            resp.json()["detail"][0]["msg"]
+            == "Opt-in notices must be served in an overlay."
+        )
+
+    def test_post_privacy_notice_opt_out_must_be_displayed_in_overlay_or_privacy_center(
+        self,
+        api_client: TestClient,
+        generate_auth_header,
+        notice_request: dict[str, Any],
+        url,
+    ):
+        """
+        Test opt out consent mechanisms should be in overlay or privacy center at the moment
+        """
+        notice_request["consent_mechanism"] = ConsentMechanism.opt_out.value
+        notice_request["displayed_in_overlay"] = False
+        notice_request["displayed_in_privacy_center"] = False
+
+        auth_header = generate_auth_header(scopes=[scopes.PRIVACY_NOTICE_CREATE])
+
+        resp = api_client.post(url, headers=auth_header, json=[notice_request])
+        assert resp.status_code == 422
+        assert (
+            resp.json()["detail"][0]["msg"]
+            == "Opt-out notices must be served in an overlay or the privacy center."
+        )
+
+    def test_post_privacy_notice_notice_only_must_be_displayed_in_overlay(
+        self,
+        api_client: TestClient,
+        generate_auth_header,
+        notice_request: dict[str, Any],
+        url,
+    ):
+        """
+        Test opt out consent mechanisms should be in overlay or privacy center at the moment
+        """
+        notice_request["consent_mechanism"] = ConsentMechanism.notice_only.value
+        notice_request["displayed_in_overlay"] = False
+        notice_request["displayed_in_privacy_center"] = False
+
+        auth_header = generate_auth_header(scopes=[scopes.PRIVACY_NOTICE_CREATE])
+
+        resp = api_client.post(url, headers=auth_header, json=[notice_request])
+        assert resp.status_code == 422
+        assert (
+            resp.json()["detail"][0]["msg"]
+            == "Notice-only notices must be served in an overlay."
+        )
+
     def test_post_privacy_notice(
         self,
         api_client: TestClient,
@@ -1558,6 +1694,7 @@ class TestPatchPrivacyNotices:
             ],
             "consent_mechanism": ConsentMechanism.opt_out.value,
             "data_uses": ["advertising"],
+            "displayed_in_overlay": True,
         }
 
     @pytest.fixture(scope="function")
@@ -1915,6 +2052,85 @@ class TestPatchPrivacyNotices:
             ],
         )
         assert resp.status_code == 422
+
+    def test_patch_privacy_notice_opt_in_must_be_displayed_in_overlay(
+        self,
+        api_client: TestClient,
+        generate_auth_header,
+        patch_privacy_notice_payload: dict[str, Any],
+        url,
+    ):
+        """
+        Test patching a single privacy notice is invalid if opt in notice is not displayed in overlay
+        """
+        auth_header = generate_auth_header(scopes=[scopes.PRIVACY_NOTICE_UPDATE])
+        patch_privacy_notice_payload[
+            "consent_mechanism"
+        ] = ConsentMechanism.opt_in.value
+        patch_privacy_notice_payload["displayed_in_overlay"] = False
+        patch_privacy_notice_payload["displayed_in_privacy_center"] = True
+
+        resp = api_client.patch(
+            url, headers=auth_header, json=[patch_privacy_notice_payload]
+        )
+        assert resp.status_code == 422
+        assert (
+            resp.json()["detail"][0]["msg"]
+            == "Opt-in notices must be served in an overlay."
+        )
+
+    def test_patch_privacy_notice_opt_out_must_be_displayed_in_overlay_or_pc(
+        self,
+        api_client: TestClient,
+        generate_auth_header,
+        patch_privacy_notice_payload: dict[str, Any],
+        url,
+    ):
+        """
+        Test patching a single privacy notice is invalid if opt out notice is not displayed in overlay or privacy center
+        """
+        auth_header = generate_auth_header(scopes=[scopes.PRIVACY_NOTICE_UPDATE])
+        assert (
+            patch_privacy_notice_payload["consent_mechanism"]
+            == ConsentMechanism.opt_out.value
+        )
+        patch_privacy_notice_payload["displayed_in_overlay"] = False
+        patch_privacy_notice_payload["displayed_in_privacy_center"] = False
+
+        resp = api_client.patch(
+            url, headers=auth_header, json=[patch_privacy_notice_payload]
+        )
+        assert resp.status_code == 422
+        assert (
+            resp.json()["detail"][0]["msg"]
+            == "Opt-out notices must be served in an overlay or the privacy center."
+        )
+
+    def test_patch_privacy_notice_notice_only_must_be_displayed_in_overlay(
+        self,
+        api_client: TestClient,
+        generate_auth_header,
+        patch_privacy_notice_payload: dict[str, Any],
+        url,
+    ):
+        """
+        Test patching a single privacy notice is invalid if notice only is not displayed in overlay
+        """
+        auth_header = generate_auth_header(scopes=[scopes.PRIVACY_NOTICE_UPDATE])
+        patch_privacy_notice_payload[
+            "consent_mechanism"
+        ] = ConsentMechanism.notice_only.value
+        patch_privacy_notice_payload["displayed_in_overlay"] = False
+        patch_privacy_notice_payload["displayed_in_privacy_center"] = True
+
+        resp = api_client.patch(
+            url, headers=auth_header, json=[patch_privacy_notice_payload]
+        )
+        assert resp.status_code == 422
+        assert (
+            resp.json()["detail"][0]["msg"]
+            == "Notice-only notices must be served in an overlay."
+        )
 
     def test_patch_privacy_notice(
         self,